Revert "Session Environments, and Editable Sessions (#276)"
This reverts commit 666a4369cea553c92e5761c0178d950c3e0787e1.
