+++ src/app/Http/Controllers/Sessions/SessionController.php
@@ -3,23 +3,17 @@
 namespace App\Http\Controllers\Sessions;
 
 use App\Http\Controllers\Controller;
-use App\Http\Resources\ComponentResource;
 use App\Http\Resources\SessionResource;
 use App\Http\Resources\TestRunResource;
 use App\Http\Resources\UseCaseResource;
-use App\Models\GroupEnvironment;
 use App\Models\Session;
-use App\Models\TestCase;
 use App\Models\UseCase;
 use Exception;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use Inertia\Inertia;
 use Inertia\Response;
-use Throwable;
 
 class SessionController extends Controller
 {
@@ -40,12 +34,12 @@ class SessionController extends Controller
             'sessions' => SessionResource::collection(
                 Session::whereHas('owner', function (Builder $query) {
                     $query
-                        ->whereKey(auth()->user()->getAuthIdentifier())
+                        ->whereKey(auth()->user())
                         ->orWhereHas('groups', function (Builder $query) {
                             $query->whereHas('users', function (
                                 Builder $query
                             ) {
-                                $query->whereKey(auth()->user()->getAuthIdentifier());
+                                $query->whereKey(auth()->user());
                             });
                         });
                 })
@@ -66,10 +60,10 @@ class SessionController extends Controller
                 Builder $query
             ) {
                 $query
-                    ->whereKey(auth()->user()->getAuthIdentifier())
+                    ->whereKey(auth()->user())
                     ->orWhereHas('groups', function (Builder $query) {
                         $query->whereHas('users', function (Builder $query) {
-                            $query->whereKey(auth()->user()->getAuthIdentifier());
+                            $query->whereKey(auth()->user());
                         });
                     });
             })->count(),
@@ -107,150 +101,6 @@ class SessionController extends Controller
                     ->paginate()
             ),
         ]);
-    }
-
-    /**
-     * @param Session $session
-     * @return Response
-     * @throws AuthorizationException
-     */
-    public function edit(Session $session)
-    {
-        $this->authorize('update', $session);
-        $component = $session->components()->firstOrFail();
-
-        return Inertia::render('sessions/edit', [
-            'session' => (new SessionResource(
-                $session->load([
-                    'testCases' => function ($query) {
-                        return $query->with(['useCase', 'lastTestRun']);
-                    },
-                    'groupEnvironment',
-                ])
-            ))->resolve(),
-            'component' => (new ComponentResource($component))->resolve(),
-            'useCases' => UseCaseResource::collection(
-                UseCase::with([
-                    'testCases' => function ($query) use ($component) {
-                        $query
-                            ->whereHas('components', function ($query) use ($component) {
-                                $query->whereKey($component->getKey());
-                            })
-                            ->where('public', true)
-                            ->orWhereHas('owner', function ($query) {
-                                $query->whereKey(auth()->user()->getAuthIdentifier());
-                            })
-                            ->orWhereHas('groups', function ($query) {
-                                $query->whereHas('users', function ($query) {
-                                    $query->whereKey(auth()->user()->getAuthIdentifier());
-                                });
-                            })
-                            ->when(
-                                auth()
-                                    ->user()
-                                    ->can('viewAnyPrivate', TestCase::class),
-                                function ($query) {
-                                    $query->orWhere('public', false);
-                                }
-                            );
-                    },
-                ])
-                    ->whereHas('testCases', function ($query) use ($component) {
-                        $query
-                            ->whereHas('components', function ($query) use ($component) {
-                                $query->whereKey($component->getKey());
-                            })
-                            ->when(
-                                !auth()
-                                    ->user()
-                                    ->can('viewAny', TestCase::class),
-                                function ($query) {
-                                    $query->where('public', true);
-                                }
-                            );
-                    })
-                    ->get()
-            ),
-            'hasGroupEnvironments' => GroupEnvironment::whereHas('group', function (Builder $query) {
-                $query->whereHas('users', function (Builder $query) {
-                    $query->whereKey(auth()->user()->getAuthIdentifier());
-                });
-            })->exists(),
-        ]);
-    }
-
-    /**
-     * @param Session $session
-     * @param Request $request
-     * @return RedirectResponse
-     * @throws AuthorizationException
-     * @throws Throwable
-     */
-    public function update(Session $session, Request $request)
-    {
-        $this->authorize('update', $session);
-        $request->validate([
-            'name' => ['required', 'string', 'max:255'],
-            'description' => ['string', 'nullable'],
-            'group_environment_id' => ['nullable', 'exists:group_environments,id'],
-            'environments.*.name' => ['required'],
-            'environments.*.value' => ['required'],
-            'component_id' => ['required', 'exists:components,id'],
-            'component_base_url' => ['required', 'url', 'max:255'],
-            'test_cases' => ['required', 'array', 'exists:test_cases,id'],
-        ]);
-
-        try {
-            $session = DB::transaction(function () use ($session, $request) {
-                $session->update($request->input());
-
-                $session->components()
-                    ->updateExistingPivot($request->input('component_id'), [
-                        'base_url' => $request->input('component_base_url'),
-                    ]);
-
-                $session->testCasesWithSoftDeletes()
-                    ->whereKey($request->input('test_cases'))
-                    ->each(function ($testCase) {
-                        $testCase->pivot->update(['deleted_at' => null]);
-                    });
-
-                $session->testCasesWithSoftDeletes()
-                    ->whereKeyNot($request->input('test_cases'))
-                    ->whereHas('testRunsWithSoftDeletesTestCases', function ($query) use ($session) {
-                        $query->where('session_id', $session->getKey());
-                    })
-                    ->each(function ($testCase) {
-                        $testCase->pivot->update(['deleted_at' => $testCase->fromDateTime($testCase->freshTimestamp())]);
-                    });
-
-                $session->testCasesWithSoftDeletes()
-                    ->whereKeyNot($request->input('test_cases'))
-                    ->whereDoesntHave('testRunsWithSoftDeletesTestCases', function ($query) use ($session) {
-                        $query->where('session_id', $session->getKey());
-                    })
-                    ->each(function ($testCase) use ($session) {
-                        $testCase->pivot->delete();
-                    });
-
-                $session->testCasesWithSoftDeletes()
-                    ->attach(
-                        collect($request->input('test_cases'))
-                            ->diff($session->testCasesWithSoftDeletes()->pluck('id'))
-                            ->all()
-                    );
-
-                return $session;
-            });
-
-            return redirect()
-                ->route('sessions.show', $session)
-                ->with('success', __('Session updated successfully'));
-        } catch (Throwable $e) {
-            return redirect()
-                ->back()
-                ->with('error', $e->getMessage());
-        }
     }
 
     /**
