+++ src/app/Http/Controllers/Testing/SimulatorController.php
@@ -64,9 +64,6 @@ class SimulatorController extends Controller
             ->whereHas('target', function ($query) use ($connection) {
                 $query->whereKey($connection->getKey());
             })
-            ->whereHas('testCase', function ($query) {
-                $query->where('deleted_at', null);
-            })
             ->offset(
                 $testRun
                     ->testResults()
