+++ src/app/Http/Controllers/Testing/SutController.php
@@ -85,9 +85,6 @@ class SutController extends Controller
             ->whereHas('target', function ($query) use ($connection) {
                 $query->whereKey($connection->getKey());
             })
-            ->whereHas('testCase', function ($query) {
-                $query->where('deleted_at', null);
-            })
             ->where(function ($query) {
                 $query
                     ->where(function ($query) {
