+++ src/app/Http/Resources/ComponentResource.php
@@ -16,9 +16,7 @@ class ComponentResource extends JsonResource
             'id' => $this->id,
             'uuid' => $this->uuid,
             'name' => $this->name,
-            'base_url' => $this->whenPivotLoaded('session_components', function () {
-                return $this->pivot->base_url;
-            }, $this->base_url),
+            'base_url' => $this->base_url,
             'position' => $this->position,
             'connections' => static::collection(
                 $this->whenLoaded('connections')
