--- src/app/Http/Resources/GroupEnvironmentResource.php
@@ -1,29 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace App\Http\Resources;
-
-use Illuminate\Http\Resources\Json\JsonResource;
-
-class GroupEnvironmentResource extends JsonResource
-{
-    /**
-     * @param  \Illuminate\Http\Request  $request
-     * @return array
-     */
-    public function toArray($request)
-    {
-        return [
-            'id' => $this->id,
-            'name' => $this->name,
-            'variables' => $this->variables,
-            'can' => [
-                'update' => auth()
-                    ->user()
-                    ->can('update', $this->resource),
-                'delete' => auth()
-                    ->user()
-                    ->can('delete', $this->resource),
-            ],
-        ];
-    }
-}
