+++ src/app/Models/UseCase.php
@@ -4,11 +4,9 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Facades\DB;
 
 /**
  * @mixin \Eloquent
- * @method static withTestCasesOfSession(Session $session)
  */
 class UseCase extends Model
 {
@@ -46,22 +44,16 @@ class UseCase extends Model
                                 $query->where('session_id', $session->getKey());
                             },
                         ])
-                        ->whereExists(function ($query) use ($session) {
-                            $query->select(DB::raw(1))
-                                ->from('session_test_cases')
-                                ->where('session_test_cases.deleted_at', null)
-                                ->where('session_test_cases.session_id', $session->getKey())
-                                ->whereColumn('session_test_cases.test_case_id', 'test_cases.id');
+                        ->whereHas('sessions', function ($query) use (
+                            $session
+                        ) {
+                            $query->whereKey($session->getKey());
                         });
                 },
             ])
             ->whereHas('testCases', function ($query) use ($session) {
-                $query->whereExists(function ($query) use ($session) {
-                    $query->select(DB::raw(1))
-                        ->from('session_test_cases')
-                        ->where('session_test_cases.deleted_at', null)
-                        ->where('session_test_cases.session_id', $session->getKey())
-                        ->whereColumn('session_test_cases.test_case_id', 'test_cases.id');
+                $query->whereHas('sessions', function ($query) use ($session) {
+                    $query->whereKey($session->getKey());
                 });
             });
     }
