+++ src/composer.json
@@ -4,15 +4,14 @@
     "description": "GSMA ITP Inclusive Tech Lab",
     "license": "MIT",
     "require": {
-        "php": "^7.3.0",
+        "php": "^7.2.5",
         "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
+        "fruitcake/laravel-cors": "^1.0",
+        "guzzlehttp/guzzle": "^6.5",
         "inertiajs/inertia-laravel": "^0.2.5",
-        "laravel/framework": "^8.0",
-        "laravel/legacy-factories": "^1.0",
+        "laravel/framework": "^7.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^3.0",
+        "laravel/ui": "^2.0",
         "league/openapi-psr7-validator": "^0.7.0",
         "nyholm/psr7": "^1.2",
         "phpunit/phpunit": "^9.1",
@@ -27,7 +26,7 @@
         "fzaninotto/faker": "^1.9.1",
         "laravel/dusk": "^6.1",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^5.0"
+        "nunomaduro/collision": "^4.1"
     },
     "config": {
         "optimize-autoloader": true,
