+++ src/resources/js/layouts/sessions/main.vue
@@ -1,30 +1,96 @@
 <template>
-    <layout :session="session">
-        <div class="col-3 mt-3 pr-0">
-            <div class="card mb-0">
-                <div class="card-header px-3">
-                    <h3 class="card-title">
-                        Select use cases
-                    </h3>
-                </div>
-                <div class="card-body p-0">
-                    <ul class="list-unstyled">
-                        <li
-                            v-for="useCase in useCases.data"
-                            v-bind:key="useCase.id"
+    <layout>
+        <div class="row">
+            <div class="col-md-12">
+                <div class="row border-bottom pb-4 align-items-stretch">
+                    <div class="col-6 d-flex flex-column">
+                        <div class="page-pretitle font-weight-normal">
+                            <breadcrumb
+                                class="breadcrumb-bullets"
+                                :items="breadcrumbs"
+                            ></breadcrumb>
+                        </div>
+                        <h1 class="page-title mw-100 mr-2 text-break">
+                            <b>{{ session.name }}</b>
+                        </h1>
+                    </div>
+                    <div class="ml-auto col-1 d-flex justify-content-end">
+                        <inertia-link
+                            :href="
+                                route('sessions.message-log.index', session.id)
+                            "
+                            class="btn btn-secondary mr-2"
                         >
-                            <b
-                                class="d-block dropdown-toggle py-2 px-3 border-bottom"
-                                v-b-toggle="
+                            <icon name="list"></icon>
+                            Log
+                        </inertia-link>
+                    </div>
+                    <div class="col-2 d-flex justify-content-end">
+                        <div class="w-100">
+                            <div>
+                                Execution:
+                                <icon
+                                    name="briefcase"
+                                    v-b-tooltip.hover
+                                    title="Use Case"
+                                />
+                                <small>
+                                    {{
+                                        session.testCases
+                                            ? collect(session.testCases.data)
+                                                  .map(function (value) {
+                                                      return value.useCase.id;
+                                                  })
+                                                  .unique()
+                                                  .count()
+                                            : 0
+                                    }}
+                                </small>
+                                <icon
+                                    name="file-text"
+                                    v-b-tooltip.hover
+                                    title="Test Case"
+                                />
+                                <small>
+                                    {{
+                                        session.testCases
+                                            ? session.testCases.data.length
+                                            : 0
+                                    }}
+                                </small>
+                            </div>
+                            <session-progress
+                                :testCases="session.testCases.data"
+                            />
+                        </div>
+                    </div>
+                </div>
+                <div class="row align-items-start">
+                    <div class="col-3 mt-3 pr-0">
+                        <div class="card mb-0">
+                            <div class="card-header px-3">
+                                <h3 class="card-title">
+                                    Select use cases
+                                </h3>
+                            </div>
+                            <div class="card-body p-0">
+                                <ul class="list-unstyled">
+                                    <li
+                                        v-for="useCase in useCases.data"
+                                        v-bind:key="useCase.id"
+                                    >
+                                        <b
+                                            class="d-block dropdown-toggle py-2 px-3 border-bottom"
+                                            v-b-toggle="
                                                 `use-case-${useCase.id}`
                                             "
-                            >
-                                {{ useCase.name }}
-                            </b>
-                            <b-collapse
-                                :id="`use-case-${useCase.id}`"
-                                visible
-                                v-if="
+                                        >
+                                            {{ useCase.name }}
+                                        </b>
+                                        <b-collapse
+                                            :id="`use-case-${useCase.id}`"
+                                            visible
+                                            v-if="
                                                 collect(useCase.testCases)
                                                     .where(
                                                         'behavior',
@@ -32,9 +98,9 @@
                                                     )
                                                     .count()
                                             "
-                            >
-                                <ul class="list-unstyled">
-                                    <li>
+                                        >
+                                            <ul class="list-unstyled">
+                                                <li>
                                                     <span
                                                         class="dropdown-toggle d-block pl-4 pr-3 py-2 font-weight-medium border-bottom"
                                                         v-b-toggle="
@@ -43,15 +109,15 @@
                                                     >
                                                         Happy flow
                                                     </span>
-                                        <b-collapse
-                                            :id="`positive-test-cases-${useCase.id}`"
-                                            visible
-                                        >
-                                            <ul
-                                                class="list-unstyled"
-                                            >
-                                                <li
-                                                    v-for="positiveTestCase in collect(
+                                                    <b-collapse
+                                                        :id="`positive-test-cases-${useCase.id}`"
+                                                        visible
+                                                    >
+                                                        <ul
+                                                            class="list-unstyled"
+                                                        >
+                                                            <li
+                                                                v-for="positiveTestCase in collect(
                                                                     useCase.testCases
                                                                 )
                                                                     .where(
@@ -59,20 +125,20 @@
                                                                         'positive'
                                                                     )
                                                                     .all()"
-                                                    v-bind:key="
+                                                                v-bind:key="
                                                                     positiveTestCase.id
                                                                 "
-                                                    class="list-group-item-action border-bottom"
-                                                    v-bind:class="{
+                                                                class="list-group-item-action border-bottom"
+                                                                v-bind:class="{
                                                                     'bg-body':
                                                                         testCase !==
                                                                             undefined &&
                                                                         positiveTestCase.id ===
                                                                             testCase.id,
                                                                 }"
-                                                >
-                                                    <inertia-link
-                                                        :href="
+                                                            >
+                                                                <inertia-link
+                                                                    :href="
                                                                         route(
                                                                             'sessions.test-cases.show',
                                                                             [
@@ -81,16 +147,16 @@
                                                                             ]
                                                                         )
                                                                     "
-                                                        class="d-flex justify-content-between align-items-center pl-5 pr-4 py-2"
-                                                    >
+                                                                    class="d-flex justify-content-between align-items-center pl-5 pr-4 py-2"
+                                                                >
                                                                     <span>
                                                                         {{
                                                                             positiveTestCase.name
                                                                         }}
                                                                     </span>
 
-                                                        <span
-                                                            v-if="
+                                                                    <span
+                                                                        v-if="
                                                                             positiveTestCase.lastTestRun &&
                                                                             positiveTestCase
                                                                                 .lastTestRun
@@ -99,10 +165,10 @@
                                                                                 .lastTestRun
                                                                                 .successful
                                                                         "
-                                                            class="flex-shrink-0 mr-0 ml-1 badge bg-success"
-                                                        ></span>
-                                                        <span
-                                                            v-else-if="
+                                                                        class="flex-shrink-0 mr-0 ml-1 badge bg-success"
+                                                                    ></span>
+                                                                    <span
+                                                                        v-else-if="
                                                                             positiveTestCase.lastTestRun &&
                                                                             positiveTestCase
                                                                                 .lastTestRun
@@ -111,23 +177,23 @@
                                                                                 .lastTestRun
                                                                                 .successful
                                                                         "
-                                                            class="flex-shrink-0 mr-0 ml-1 badge bg-danger"
-                                                        ></span>
-                                                        <span
-                                                            v-else
-                                                            class="flex-shrink-0 mr-0 ml-1 badge bg-secondary"
-                                                        ></span>
-                                                    </inertia-link>
+                                                                        class="flex-shrink-0 mr-0 ml-1 badge bg-danger"
+                                                                    ></span>
+                                                                    <span
+                                                                        v-else
+                                                                        class="flex-shrink-0 mr-0 ml-1 badge bg-secondary"
+                                                                    ></span>
+                                                                </inertia-link>
+                                                            </li>
+                                                        </ul>
+                                                    </b-collapse>
                                                 </li>
                                             </ul>
                                         </b-collapse>
-                                    </li>
-                                </ul>
-                            </b-collapse>
-                            <b-collapse
-                                :id="`use-case-${useCase.id}`"
-                                visible
-                                v-if="
+                                        <b-collapse
+                                            :id="`use-case-${useCase.id}`"
+                                            visible
+                                            v-if="
                                                 collect(useCase.testCases)
                                                     .where(
                                                         'behavior',
@@ -135,9 +201,9 @@
                                                     )
                                                     .count()
                                             "
-                            >
-                                <ul class="list-unstyled">
-                                    <li>
+                                        >
+                                            <ul class="list-unstyled">
+                                                <li>
                                                     <span
                                                         class="dropdown-toggle d-block pl-4 pr-3 py-2 font-weight-medium border-bottom"
                                                         v-b-toggle="
@@ -146,15 +212,15 @@
                                                     >
                                                         Unhappy flow
                                                     </span>
-                                        <b-collapse
-                                            :id="`negative-test-cases-${useCase.id}`"
-                                            visible
-                                        >
-                                            <ul
-                                                class="list-unstyled"
-                                            >
-                                                <li
-                                                    v-for="negativeTestCase in collect(
+                                                    <b-collapse
+                                                        :id="`negative-test-cases-${useCase.id}`"
+                                                        visible
+                                                    >
+                                                        <ul
+                                                            class="list-unstyled"
+                                                        >
+                                                            <li
+                                                                v-for="negativeTestCase in collect(
                                                                     useCase.testCases
                                                                 )
                                                                     .where(
@@ -162,20 +228,20 @@
                                                                         'negative'
                                                                     )
                                                                     .all()"
-                                                    v-bind:key="
+                                                                v-bind:key="
                                                                     negativeTestCase.id
                                                                 "
-                                                    class="list-group-item-action border-bottom"
-                                                    v-bind:class="{
+                                                                class="list-group-item-action border-bottom"
+                                                                v-bind:class="{
                                                                     'bg-body':
                                                                         testCase !==
                                                                             undefined &&
                                                                         negativeTestCase.id ===
                                                                             testCase.id,
                                                                 }"
-                                                >
-                                                    <inertia-link
-                                                        :href="
+                                                            >
+                                                                <inertia-link
+                                                                    :href="
                                                                         route(
                                                                             'sessions.test-cases.show',
                                                                             [
@@ -184,16 +250,16 @@
                                                                             ]
                                                                         )
                                                                     "
-                                                        class="d-flex justify-content-between align-items-center pl-5 pr-4 py-2"
-                                                    >
+                                                                    class="d-flex justify-content-between align-items-center pl-5 pr-4 py-2"
+                                                                >
                                                                     <span>
                                                                         {{
                                                                             negativeTestCase.name
                                                                         }}
                                                                     </span>
 
-                                                        <span
-                                                            v-if="
+                                                                    <span
+                                                                        v-if="
                                                                             negativeTestCase.lastTestRun &&
                                                                             negativeTestCase
                                                                                 .lastTestRun
@@ -202,10 +268,10 @@
                                                                                 .lastTestRun
                                                                                 .successful
                                                                         "
-                                                            class="flex-shrink-0 mr-0 ml-1 badge bg-success"
-                                                        ></span>
-                                                        <span
-                                                            v-else-if="
+                                                                        class="flex-shrink-0 mr-0 ml-1 badge bg-success"
+                                                                    ></span>
+                                                                    <span
+                                                                        v-else-if="
                                                                             negativeTestCase.lastTestRun &&
                                                                             negativeTestCase
                                                                                 .lastTestRun
@@ -214,28 +280,31 @@
                                                                                 .lastTestRun
                                                                                 .successful
                                                                         "
-                                                            class="flex-shrink-0 mr-0 ml-1 badge bg-danger"
-                                                        ></span>
-                                                        <span
-                                                            v-else
-                                                            class="flex-shrink-0 mr-0 ml-1 badge bg-secondary"
-                                                        ></span>
-                                                    </inertia-link>
+                                                                        class="flex-shrink-0 mr-0 ml-1 badge bg-danger"
+                                                                    ></span>
+                                                                    <span
+                                                                        v-else
+                                                                        class="flex-shrink-0 mr-0 ml-1 badge bg-secondary"
+                                                                    ></span>
+                                                                </inertia-link>
+                                                            </li>
+                                                        </ul>
+                                                    </b-collapse>
                                                 </li>
                                             </ul>
                                         </b-collapse>
                                     </li>
                                 </ul>
-                            </b-collapse>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-        <div class="col-9 mt-3">
-            <div class="row">
-                <div class="col">
-                    <slot></slot>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="col-9 mt-3">
+                        <div class="row">
+                            <div class="col">
+                                <slot></slot>
+                            </div>
+                        </div>
+                    </div>
                 </div>
             </div>
         </div>
@@ -243,11 +312,18 @@
 </template>
 
 <script>
-import Layout from '@/layouts/sessions/app';
+import Layout from '@/layouts/main';
+import SessionProgress from '@/components/sessions/progress';
 
 export default {
     components: {
         Layout,
+        SessionProgress,
+    },
+    metaInfo() {
+        return {
+            title: this.session.name,
+        };
     },
     props: {
         session: {
@@ -261,6 +337,16 @@ export default {
         useCases: {
             type: Object,
             required: true,
+        },
+        breadcrumbs: {
+            type: Array,
+            required: false,
+            default: function () {
+                return [
+                    { name: 'Sessions', url: route('sessions.index') },
+                    { name: this.session.name },
+                ];
+            },
         },
     },
 };
