+++ src/resources/js/pages/groups/invite.vue
@@ -54,7 +54,7 @@
                         </div>
                         <div class="card-footer text-right">
                             <inertia-link
-                                :href="route('groups.users.index', group.id)"
+                                :href="route('groups.show', group.id)"
                                 class="btn btn-link"
                             >
                                 Cancel
@@ -122,7 +122,7 @@ export default {
         },
         loadUserList(query = '') {
             axios
-                .get(route('groups.users.candidates', this.group), {
+                .get(route('groups.user-candidates', this.group), {
                     params: { q: query },
                 })
                 .then((result) => {
