+++ src/resources/js/pages/sessions/test-cases/show.vue
@@ -1,6 +1,7 @@
 <template>
     <layout
         :session="session"
+        :useCases="useCases"
         :test-case="testCase"
         :testStepFirstSource="testStepFirstSource"
     >
@@ -128,6 +129,10 @@ export default {
     },
     props: {
         session: {
+            type: Object,
+            required: true,
+        },
+        useCases: {
             type: Object,
             required: true,
         },
