+++ src/resources/js/pages/sessions/test-runs/show.vue
@@ -2,6 +2,7 @@
     <layout
         :session="session"
         :testCase="testCase"
+        :useCases="useCases"
         :testStepFirstSource="testStepFirstSource"
     >
         <div class="card">
@@ -717,6 +718,10 @@ export default {
     },
     props: {
         session: {
+            type: Object,
+            required: true,
+        },
+        useCases: {
             type: Object,
             required: true,
         },
