+++ src/resources/js/pages/sessions/test-steps/flow.vue
@@ -2,6 +2,7 @@
     <layout
         :session="session"
         :testCase="testCase"
+        :useCases="useCases"
         :testStepFirstSource="testStepFirstSource"
     >
         <div class="card mb-0">
@@ -46,6 +47,10 @@ export default {
             required: true,
         },
         testCase: {
+            type: Object,
+            required: true,
+        },
+        useCases: {
             type: Object,
             required: true,
         },
